Third iteration

+++ Project1/mysite/food/models.py
@@ -4,8 +4,9 @@ from django.db import models
 class Item(models.Model):
     def __str__(self):
         return self.item_name
+               
 
     item_name = models.CharField(max_length= 200)
     item_desc = models.CharField(max_length= 200)
-    item_price = models.IntegerField()
+    item_price = models.FloatField()
     item_image = models.CharField(max_length=500, default= 'https://external-content.duckduckgo.com/iu/?u=http%3A%2F%2Fwww.ukvisitorguide.cn%2Fwp-content%2Fuploads%2F2015%2F11%2FFood-placeholder.jpg&f=1&nofb=1')
